Move PlayServicesChecker & FCMTokenRepository to the common-android module
