+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/core/injection/AndroidAppComponentImpl.kt
@@ -14,8 +14,6 @@ import org.hyperskill.app.android.notification.local.injection.PlatformLocalNoti
 import org.hyperskill.app.android.notification.local.injection.PlatformLocalNotificationComponentImpl
 import org.hyperskill.app.android.notification.remote.injection.AndroidPlatformPushNotificationComponent
 import org.hyperskill.app.android.notification.remote.injection.AndroidPlatformPushNotificationsComponentImpl
-import org.hyperskill.app.android.notification.remote.injection.AndroidPlatformPushNotificationsPlatformDataComponent
-import org.hyperskill.app.android.play_services.injection.PlayServicesCheckerComponentImpl
 import org.hyperskill.app.android.sentry.domain.model.manager.SentryManagerImpl
 import org.hyperskill.app.core.domain.BuildVariant
 import org.hyperskill.app.core.injection.CommonAndroidAppGraphImpl
@@ -24,8 +22,6 @@ import org.hyperskill.app.core.injection.CommonComponentImpl
 import org.hyperskill.app.core.remote.UserAgentInfo
 import org.hyperskill.app.main.injection.PlatformMainComponent
 import org.hyperskill.app.main.injection.PlatformMainComponentImpl
-import org.hyperskill.app.notification.remote.injection.PlatformPushNotificationsDataComponent
-import org.hyperskill.app.play_services.injection.PlayServicesCheckerComponent
 import org.hyperskill.app.sentry.injection.SentryComponent
 import org.hyperskill.app.sentry.injection.SentryComponentImpl
 
@@ -62,14 +58,6 @@ class AndroidAppComponentImpl(
             commonComponent = commonComponent,
             platformLocalNotificationComponent = platformLocalNotificationComponent,
             analyticInteractor = analyticComponent.analyticInteractor
-        )
-
-    override fun buildPlayServicesCheckerComponent(): PlayServicesCheckerComponent =
-        PlayServicesCheckerComponentImpl(context, sentryComponent)
-
-    override fun buildPlatformPushNotificationsDataComponent(): PlatformPushNotificationsDataComponent =
-        AndroidPlatformPushNotificationsPlatformDataComponent(
-            playServicesCheckerComponent = buildPlayServicesCheckerComponent(),
         )
 
     /**
