+++ shared/src/androidMain/kotlin/org/hyperskill/app/play_services/injection/PlayServicesCheckerComponentImpl.kt
@@ -1,9 +1,8 @@
-package org.hyperskill.app.android.play_services.injection
+package org.hyperskill.app.play_services.injection
 
 import android.content.Context
-import org.hyperskill.app.android.play_services.domain.model.PlayServicesCheckerImpl
+import org.hyperskill.app.play_services.data.PlayServicesCheckerImpl
 import org.hyperskill.app.play_services.domain.PlayServicesChecker
-import org.hyperskill.app.play_services.injection.PlayServicesCheckerComponent
 import org.hyperskill.app.sentry.injection.SentryComponent
 
 class PlayServicesCheckerComponentImpl(
